// src/enums/path.ts
export enum Path {
    Root = '/',
    AppSettings = '/app-settings',
    AddCategory = '/add-category',
    PrivacyPolicy = '/privacy-policy',
    User = '/user',
    Users = '/users',
    ChangeStatus = '/change-status',
    ChangeRights = '/change-rights',
    Register = '/register',
    Login = '/login',
    Logout = '/logout',
    Auth = '/auth',
    ChangeAvatar = '/change-avatar',
    CreateReview = '/create-review',
    EditReview = '/edit-review',
    Reviews = '/reviews',
    Review = '/review',
    Like = '/like',
    Star = '/star',
    Tag = '/tag',
    Social = '/social',
    Delete = '/delete',
    CreateComment = '/create-comment',
    Id = ':id'
}